backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.1.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
